DOC-Comparison_1/requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-openai
openai
pypdf2
python-docx
faiss-cpu